requirements.txt
flask-swagger-ui
flask-restplus
attr
Flask==0.12.1
Flask-RESTful==0.3.5
pytest==3.0.7
elasticsearch6>=6.0.0,<7.0.0
